requirements.txt
PyQt5
openai-whisper
sounddevice
numpy